Mypy fixes
